+++ packages/ui/src/modals/TransferModal/TransactionFailureModal.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import { ButtonPrimaryMedium } from '../../components/buttons/Buttons'
 import { ArrowDownIcon, Icon } from '../../components/icons/ArrowDownIcon'
 import { SuccessIcon } from '../../components/icons/SuccessIcon'
-import { ModalHeader, ResultModal, SuccessModalBody } from '../../components/modal'
+import { ModalHeader, ResultModal, SuccessModalBody } from '../../components/Modal'
 import { TokenValue } from '../../components/typography'
 import { Account } from '../../hooks/types'
 import { TransactionInfoLabel } from '../common'
