+++ packages/ui/src/modals/TransferModal/TransactionSuccessModal.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { AccountInfo } from '../../components/AccountInfo'
 import { ArrowDownExpandedIcon } from '../../components/icons/ArrowDownExpandedIcon'
 import { SuccessIcon } from '../../components/icons/SuccessIcon'
-import { ModalHeader, ResultModal, SuccessModalBody } from '../../components/modal'
+import { ModalHeader, ResultModal, SuccessModalBody } from '../../components/Modal'
 import { TokenValue } from '../../components/typography'
 import { Account } from '../../hooks/types'
 import { useBalance } from '../../hooks/useBalance'
