+++ packages/ui/src/modals/TransferModal/TransferDetailsModal.tsx
@@ -3,7 +3,7 @@ import React, { ReactElement, useState } from 'react'
 import styled from 'styled-components'
 import { AccountInfo } from '../../components/AccountInfo'
 import { ButtonPrimaryMedium, ButtonSecondarySmall } from '../../components/buttons/Buttons'
-import { Modal, ModalBody, ModalFooter, ModalHeader } from '../../components/modal'
+import { Modal, ModalBody, ModalFooter, ModalHeader } from '../../components/Modal'
 import { TokenValue } from '../../components/typography'
 import { SelectAccount } from '../../components/selects/AccountSelectTemplate/SelectAccount'
 import { BorderRad, Colors } from '../../constants'
